+++ CHANGELOG.rst
@@ -2,6 +2,13 @@
 Changelog for package plotjuggler
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.0.7 (2017-05-12)
+------------------
+* the list of topics in the Dialog will be automatically updated
+* bug fix
+* fixed some issues with the installation
+* Contributors: Davide Faconti
+
 1.0.5 (2017-05-07)
 ------------------
 * fixed an issue with ROS during destruction